backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==2.3.0
python-dotenv==1.0.0
openai==1.8.5
httpx==0.24.1
